+++ History.md
@@ -1,3 +1,8 @@
+0.1.1 / 2012-10-25
+==================
+
+* Added enhanced link attribution ability for Google Analytics
+* Added site speed sample rate for Google Analytics
 
 0.1.0 / 2012-10-11 
 ==================